pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "setuptools_scm[toml]>=3.4.3",
    "wheel",
    "pyproject-flake8",
]
build-backend = "setuptools.build_meta"

[project]
name = "sgnl"
dynamic = ["version"]
dependencies = [
    # External packages
    "pandas",
    # LIGO dependencies
    "ligo-gracedb==2.14.1",
    "lalsuite>=7.26",
    "igwn-ligolw<2.1.0",
    "igwn-segments",
    # SGN dependencies
    "sgn",
    "sgn-ts",
    "sgn-ligo",
    "flask",# FIXME remove this once we move the gracedb mock server out of here.
    "ligo-scald",
]

requires-python = ">=3.11"
authors = [
    { name = "Chad Hanna", email = "crh184@psu.edu" },
    { name = "Yun-Jing Huang", email = "yzh5436@psu.edu" },
    { name = "James Kennington", email = "jmk7376@psu.edu" },
]
description = "Inspiral Search Pipeline using SGN Framework"
readme = "README.md"
license = { text = "MPL-2.0" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://git.ligo.org/greg/sgnl"
Documentation = "https://docs.ligo.org/greg/sgnl"
Issues = "https://git.ligo.org/greg/sgnl/issues"

[project.scripts]
sgnl-add-segments = "sgnl.bin.add_segments:main"
sgnl-assign-far = "sgnl.bin.assign_far:main"
sgnl-assign-likelihood = "sgnl.bin.assign_likelihood:main"
sgnl-create-prior-diststats = "sgnl.bin.create_prior_diststats:main"
sgnl-dagger = "sgnl.bin.dagger:main"
sgnl-extinct-bin = "sgnl.bin.extinct_bin:main"
sgnl-flopulator = "sgnl.bin.flopulator:main"
sgnl-inspiral = "sgnl.bin.inspiral:main"
sgnl-inspiral-bank-splitter = "sgnl.bin.inspiral_bank_splitter:main"
sgnl-inspiral-set-svdbin-option = "sgnl.bin.inspiral_set_svdbin_option:main"
sgnl-inspiral-svd-bank = "sgnl.bin.inspiral_svd_bank:main"
sgnl-ll-dagger = "sgnl.bin.ll_dagger:main"
sgnl-ll-dq = "sgnl.bin.ll_dq:main"
sgnl-ll-inspiral-event-uploader = "sgnl.bin.ll_inspiral_event_uploader:main"
sgnl-ll-inspiral-event-plotter = "sgnl.bin.ll_inspiral_event_plotter:main"
sgnl-ll-marginalize-likelihoods-online = "sgnl.bin.ll_marginalize_likelihoods_online:main"
sgnl-ll-pastro-uploader = "sgnl.bin.ll_pastro_uploader:main"
sgnl-ll-remove-counts = "sgnl.bin.ll_remove_counts:main"
sgnl-ll-set-gracedb-far-threshold = "sgnl.bin.ll_set_gracedb_far_threshold:main"
sgnl-ll-trigger-counter = "sgnl.bin.ll_trigger_counter:main"
sgnl-median-of-psds = "sgnl.bin.median_of_psds:main"
sgnl-mock-gracedb-server = "sgnl.bin.mock_gracedb_server:main"
sgnl-plot-psd = "sgnl.bin.plot_psd:main"
sgnl-query-segments = "sgnl.bin.query_segments:main"
sgnl-reference-psd = "sgnl.bin.reference_psd:main"
sgnl-results-page = "sgnl.bin.results_page:main"
sgnl-sim-page = "sgnl.bin.sim_page:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "lalsuite",
    "torch<=2.5.1",
    "gwpy",
    "sgnl-test-data @ git+https://git.ligo.org/greg/sgnl-test-data.git",
]
docs = [
    "mkdocs",
    "mkdocstrings",
    "mkdocstrings-python",
    "mkdocs-material",
    "pymdown-extensions",
    "mkdocs-autorefs<1.3.0"
]
lint = [
    "black",
    "flake8",
    "flake8-bandit",
    "flake8-black",
    "flake8-bugbear",
    "flake8-future-annotations",
    "flake8-isort",
    "flake8-logging-format",
    "flake8-pyproject",
    "isort",
    "mypy",
    "mypy-extensions",
    "types-requests",
    "types-PyYAML",
]
dev = [
    "sgnl[docs]",
    "sgnl[lint]",
    "sgnl[test]",
]
dag = [
    "ezdag",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sgnl = [
    "py.typed",
    "dags/default_config.yml",
]

[tool.setuptools_scm]
fallback_version = "v0.0.0"
write_to = "src/sgnl/_version.py"


[tool.coverage.run]
source = ["sgnl"]
omit = [
    "*/site-packages/*",
    "*/dist-packages/*",
    "*/_version.py",
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.paths]
source = [
    "src/sgnl/",
    "*/sgnl/",
]

[tool.coverage.report]
omit = ["_version.py", "tests/*", "examples/*", "setup.py", "*/__init__.py"]
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
]

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
exclude = [
    "build",
    "docs",
]

[tool.pytest.ini_options]
addopts = "-p no:warnings --cov=sgnl --cov-report=term-missing"

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",
    "S101",
    "S406",
]
per-file-ignores = [
    "__init__.py:F401",
    "_version.py:BLK100",
    "docs/conf.py:E501",
    "tests/*.py:E731,S108",
    "src/sgnl/viz.py:E402,E501",
    "src/sgnl/dags/segments.py:S310",
]
exclude = [
    "build",
    "docs",
    "_version.py",
    "examples",
]
